@@ -1,7 +1,7 @@
 import { createBrowserRouter } from "react-router-dom";
-import Transcription from "./views/transcription";
-import App from "./App";
-import UpdateTranscription from "views/updateTranscription";
+import Transcription from "@/views/transcription";
+import App from "@/App";
+import UpdateTranscription from "@/views/updateTranscription";
 
 const router = createBrowserRouter([
   {
@@ -13,11 +13,10 @@ const router = createBrowserRouter([
   }, {
     path: "/add-transcription",
     element: <UpdateTranscription />,
-  },
-  {
+  }, {
     path: "/update-transcription/:transcriptionId",
     element: <UpdateTranscription />,
-  }
+  },
 ]);
 
 export default router;
